# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScaleWise API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created — verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterPending"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address using token from verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified — user is now logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid token
        "410":
          description: Token expired
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend the email verification link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: Verification email queued
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, emailVerified, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [client, expert, admin]
        emailVerified:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [client, expert]
        inviteToken:
          type: string
          description: One-time invite token required when registering with role=expert
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    RegisterPending:
      type: object
      required: [pending, email]
      properties:
        pending:
          type: boolean
        email:
          type: string
          description: The email address where the verification link was sent
    VerifyEmailInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: The raw verification token from the email link
    ResendVerificationInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
...