# lib/api-zod/src/generated/types/registerPending.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ScaleWise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterPending {
  pending: boolean;
  /** The email address where the verification link was sent */
  email: string;
}