# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ScaleWise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyEmailInput {
  /** The raw verification token from the email link */
  token: string;
}